implement RunPlugin for nodejs
Implement RunPlugin for NodeJS. This follows a similar path as the
rest of the language runtime does, spawning a NodeJS program, which in
turn runs the plugin itself.  The run-plugin code was cribbed from
run-policy-pack, which works similarly.

Note that this means that in addition to a recent language runtime,
this also requires the plugin to reference a recent version of the
Pulumi SDK.  Since the main purpose of implementing this is supporting
a new feature that should be okay though.