implement parameterize in the NodeJS provider interface
Parameterized providers are currently only possible to use in Go.
However we want to bring this functionality to other languages as
well.  Implement this in NodeJS.

There's potentially still an API design question here.  I've split the
parameterize method up into parameterizeValue and parameterizeArgs
functions, which should be easier to use for the implementer, as
TypeScript doesn't support overloading with different types.  Curious
about peoples thoughts on that.