add lint_fix Makefile targets
Some of the linters we're using have an option to fix some of the lint
errors they are finding.  This obviously only works for the simple
cases, and nothing too complex, but can still be useful.

Since there are a bunch of different tools we're using, it's not
always easy to remember which option needs to be passed to which tool.
Add `lint_fix` Makefile targets that pass these options to our linters
where available, to help fix these issues more easily.