set mock monitor in global store
Currently when setting a mock monitor, we only set it in a local
variable.  Unfortunately it is possible for multiple @pulumi/pulumi
modules to be loaded at the same time.  This means one of them might
have the mock monitor loaded, while the other versions don't,
resulting in confusing behaviour for the user.

Fix this by setting the mock monitor in the global settings, so it's
available in all versions of the SDK.