Add support for config file in Go Automation api
This adds support for the config file flag in the Go Automation API. It
adds the option to preview, up, refresh, and destroy and also adds the
ability to edit/get a config file via the workspace.

I had to edit some local_workspace_test tests to account for a mock
change. Previously capturedArgs would only be set to the last set of
args. This is a problem to use with up as the last thing to happen is
pulumi output to be called. I changed this type to [][]string to allow
us to capture all calls made with the mock rather than the last call.