Replace splats with explict property copying for JSON data (#17770)
Fixes https://github.com/pulumi/pulumi/issues/17734.

There were a couple of places in the python automation API where we
would read in JSON data from the CLI and
[splat](https://docs.python.org/3/tutorial/controlflow.html#unpacking-argument-lists)
the data into a constructor to return that data in a typed python class.

This hits issues whenever new data is returned from the CLI that the
python class has not yet been updated to understand (see
https://github.com/pulumi/pulumi/issues/17734 for the motivating example
this time round).
While it's easy to add the new field to the class to fix the issue,
we'll just hit this again if we ever add any other new fields.

This PR removes the splat operator and just explicitly lists out the
arguments to set. This also allows the python name to differ from the
JSON name (see
https://github.com/pulumi/pulumi/pull/17735#discussion_r1840259869 for
an example where that would have been sub-optimal).