add missing permissions to the cron-test-all job (#17764)
ci.yml that we're calling from here needs these permissions. The calling
workflow needs to include them as well, otherwise the callee is not
allowed to get them. Fix that.

I think we didn't get an issue open because of this failure because the
workflow never really ran. I'll keep an eye on it for the next days
though to make sure issues are opened correctly.