pass correct working directory to RunPlugin (#17763)
RunPlugin expects the working directory of the program to be passed in
in the plugin request. This was the case until
https://github.com/pulumi/pulumi/pull/15191. In that PR we switched from
passing pwd (which is the same as `ctx.Pwd` a layer above) to
minfo.GetProgramDirectory(), which is the path from which the plugin is
executed. Fix this by passing in `info.WorkingDirectory`, which a layer
above is `ctx.Pwd`.

Add a test for this as well to make sure we're not regressing this
again.

Likely nobody noticed this, because most plugins probably don't rely on
this path, and most providers are not being run using `RunPlugin`.