Generate more accurate reproductions when fuzzing (#17737)
Fuzzed lifecycle tests work slightly differently to their handwritten
counterparts. A handwritten test will typically "start from nothing". An
initial snapshot is built from an empty state and starting program,
before subsequent operations are executed on this state to test
behaviour. In constrast, fuzzed tests create arbitrary starting
snapshots "out of thin air", before running an operation to see if a bug
can be triggered. Ideally, any starting state conjured by a fuzz test is
actually reproducible from an empty state and some combination of
operations, but it may be that this is not the case, or that the number
of operations required to reach the state is very high. In such cases,
it is handy to have the exact code the fuzz test used to hand when
reproducing and isolating behaviour. To this end, this commit extends
the `reprogen` functionality of the suite to generate this code as well
as the existing "handwritten" approximation. This should also aid in
minimising failing test cases quickly when bugs are found.