Fix merging InvokeOptions.version (#17750)
`InvokeOptions.version` was defaulting to the empty string `""` and not
`None`, see
https://github.com/pulumi/pulumi/blob/1a97d95583a491bb32ec24fe71817f1627f1cf58/sdk/python/lib/pulumi/invoke.py#L57

While working on https://github.com/pulumi/pulumi/issues/17749 I noticed
that the merge function was not expecting the empty string, and clearing
out the version on merge when it probably shouldn't.

Since this is typed as `Optional[str]`, code that uses this value
already has to deal with `None`, and we can update it to default to
`None` instead.

---------

Co-authored-by: Thomas Gummerer <t.gummerer@gmail.com>