Workaround for control codes in output from .NET 9.0 (#17783)
.NET 9.0 introduced a bug where the output from `dotnet run` can include
a progress indicator, even though the output is redirected:
https://github.com/dotnet/sdk/issues/44610

This causes .NET plugins to fail when running under .NET 9.0 because we
can't parse the port number.

As a workaround, we strip these control characters.

This PR also tightens the port parsing to handle invalid numbers.
Previously these would error in the `dialPlugin` done just after parsing
the port.