Enable go language test for l2-parameterized-resource.
This was disabled because `go mod tidy` was failing.

The failure was caused by the paramaterized provider generating a
subpackage and attempting to replace a path to the original subpackage
part instead of to the nested package.

This change introduces a GoPackageInfo field to specify subpaths to
append specifically for go sdk generation.  It also works around this
case when getting the package name or alias when this GoPackageInfo is
available without a "path" to specify it (meaning that an append was
used, indicating this is a special case of a parameterization).