pass context to git clone utilities (#17799)
Pass a context to the git clone utilities. In the future we want to be
able to use them to download plugins, which should be cancellable. Pass
the context through to allow this.

In theory this is not backwards compatible for this API, but we've
recently been allowing similar changes (e.g. in
https://github.com/pulumi/pulumi/pull/17621), so I think this is okay to
do.