Modify packages for the user instead of printing instructions when adding package.
Before we would print instructions, but with built in tools it is not
hard to modify the go.mod on behalf of the user and print out
information.

This still leaves java in a "print instruction" state since there isnt a
standard package managment system like in the other languages.