expose a PULUMI_PLUGIN_LOCATION env variable in RunPlugin
In https://github.com/pulumi/pulumi/pull/17763, we made sure the right
working directory is being passed to `RunPlugin`.  It is however
sometimes also useful to know the directory where a plugin is being
run from, e.g. to read schema files from that location.  Expose an
environment variable to allow that.