Do not generate go.mod when generating local SDKs for (parameterized) providers.
This is related to the the previous attempt #17795 where we discovered
the actual root issue wasn't the paths but that a sub package was being
generated unnecessarily (and the paths were ok).

If there are no local dependencies that indicates this is actually
needed and not a duplicate go.mod (differentiating between adding a
package and converting a project etc).