Split public and internal representation of InvokeOptions (#17818)
In preparation for allowing the DependsOn and DependsOnInputs options to
work for (Output form) invokes, we split the public and internal
representations of InvokeOptions.

This matches resourceOptions and ResourceOptions, where the public type
is a snapshot of the private type. The private parameterization field is
moved to the private invokeOptions type.