Protobuf definition for autonaming configuration (#17810)
Introduces protobuf changes required for
[Autonaming](https://github.com/pulumi/pulumi/discussions/17592)
implementation.

We add a struct field to CheckRequest. The struct contains two fields:
- Proposed name that will hold a name calculated by the engine which is
a suggested candidate for resource's autoname (in case it's needed)
- Mode of how to apply that name:
- Disable: there should be no proposed name passed, since the autonaming
is disabled for this resource
- Enforce: provider _must_ use the proposed name if no explicit name was
provider, or error otherwise
- Propose: use the provided name as a baseline but can adjust it given
provider's domain knowledge

Once this ships, we can start upgrading providers to this protocol
version and implementing autonaming provider-side.

Also, we add a `SupportsAutonamingConfiguration` flag that a provider
needs to set to `true` in `ConfigureResponse` once it declares that it
will respec autonaming configuration provided by the engine.

Part of https://github.com/pulumi/pulumi/issues/1518