fix RunPlugin for NodeJS (#17836)
RunPlugin for NodeJS is currently broken in two ways:
- The program name that's passed through is currently
req.Info.EntryPoint, which generally is always `.`. This doesn't work
since we're passing `req.Pwd` as the pwd, so we need to pass
`req.Info.ProgramDirectory` instead.
- It's not passing `req.Args` correctly, which means that the NodeJS
program won't get an engine address.

Fix those and add some tests for RunPlugin. Since we're already adding
tests, add them for Go and Python as well.