Enable unused-import rule for Python (#17842)
We have a bunch of unused imports in our Python code. We've had this
rule disabled since pylint was introduced in
https://github.com/pulumi/pulumi/pull/2146. Let's enable it to keep the
code a little tidier.

Some of the unused imports were not entirely unused, but happened to
resolve an import cycle, although this was likely more by coincidence
than by design. We break the cycle by ensuring that the `log` module
does not import any other pulumi module (at import time).