Add some examples and caveats for the Go SDK (#17865)
I've been working with our Go SDK and both of these would have helped
me.

The exact syntax for `pulumi.All` isn't obvious (and could be improved
IMO).

The caveat in `pulumi.JSONMarshal` is extremely non-obvious, and should
be called out in our documentation until it is fixed.

TODO: After this merges, comment on
https://github.com/pulumi/pulumi/issues/12460 that the caveat needs to
be removed after the issue is fixed.