[sdk/nodejs] Throw from output() on circular structures (#17852)
Note to reviewers: Review each commit separately. The first commit
contains only formatting changes (no functional changes). The actual fix
is in the second commit.

---

`pulumi.output(...)` cannot handle circular structures. Passing one to
`output()` will currently result in a "Maximum call stack exceeded"
error.

This commit adds a better error message for when a circular structure is
detected, so that users know what the actual problem is and can address
it. The error that is thrown is similar to the error thrown when passing
a circular structure to `JSON.stringify()`.

Note: It may be possible to make circular structures work with
`output()`, but it would require more time/effort to get the
implementation right and not regress the perf fix from
https://github.com/pulumi/pulumi/pull/3851.

Fixes #17778