[DRAFT] Only check the package name, not the respsitory, to match packages.
Now that we have package blocks, they take the form
"[HOSTNAME]/NAMESPACE/packageName"

This causes there to be a naming mismatch when reading from the package
block and causing failures in tests.  This should mitigate that issue,
at the cost of risk of package name conflicts accross repos.

We should be matching against "NAMESPACE/packageName" to be inline with:
https://developer.hashicorp.com/terraform/language/providers/requirements#source-addresses

So I'm going to investigate this while others weigh in.