Move `sdk/go/common/tools/path` to `pkg`
`sdk/go/common/tools/path` is home to some helper functions that relate
to creating test Go environments. Some of this functionality is unused,
and some of it is [only used
internally](https://github.com/search?q=%28org%3Apulumiverse+OR+org%3Apulumi%29+%22CreateTemporaryGoFolder%22&type=code)
and doesn't need to be part of the public API that `sdk/go/common` in
theory comprises. This commit moves it alongside its consumer in `pkg`
to try and reduce unnecessary surface in `sdk/go/common`.