Add GetRequiredPackages to the language host (#17894)
Next part of https://github.com/pulumi/pulumi/issues/17507

This adds a new `GetRequiredPackages` method to the language hosts. This
will replace `GetRequiredPlugins`. This change set adds the method, and
the required handling in the engine without making use of the new
information returned by the package form, and without updating the
language hosts to return that information.

This should continue to work by hitting all the compatibility paths.
I'll follow this change up with updating the conformance tests to check
against this, and fixing the language hosts to return the new package
information.