Use PCL packages to decide the expected program dependencies (#17900)
This updates the conformance tests to use the list of packages used by
the PCL program for the expected dependencies, rather than the list of
provider plugins assigned for the test.

This fixes the TODO that we weren't correctly dealing with parameterized
packages because the program wouldn't depend on the provider plugin
package, but the parameterized package instead.