Break out metadata management from `pkg/cmd/pulumi/util` (#17911)
`pkg/cmd/pulumi/util.go` has a lot of diverse functionality within it.
This commit starts the process of breaking this out into subpackages so
that things can be better namespaced and encapsulated, factoring out
operation metadata into a new `util` package. Later on it may be that
`util` itself is replaced with packages like `metadata`, etc. but now
let's take the smallest steps possible as we start to clean things up.