Suggest `state repair` as part of integrity panics (#17919)
When the Pulumi CLI encounters a snapshot integrity error, we emit an
error message asking the user to file an issue. We don't however give
them much advice on how to proceed. In #17445 we introduced the `state
repair` command, which can fix a large number of issues that cause
snapshot integrity failures. In this commit, we inform users of `state
repair` in the hope that it might unblock them, or offer a neater
solution than e.g. `state delete` and the like that we've previously
observed people using. As a fallback, we also mention the
`--disable-integrity-checking` flag that users have used previously, to
make it a bit more discoverable.

Fixes #4984


![image](https://github.com/user-attachments/assets/4c030ff6-8132-4836-9f7d-f8dbc51d898f)

![image](https://github.com/user-attachments/assets/f81c844f-07ea-4ca5-989b-589169a6d166)