Drop Python 3.8 from full-language-matrix-test (#17917)
In https://github.com/pulumi/pulumi/pull/17883 we bumped the minimum
version of Python 3.8. This changed some test assertions to expect
seeing 3.9 in pyproject.toml and the like, causing tests to fail when
run on 3.8. The SDK itself probably still works fine, but in any case
this should always match the min supported version.