Break out `cmd`-level backend code
Presently, we define a `LoginManager` interface in `pkg/backend` that
serves only to provide abstractions to `pkg/cmd/pulumi`. This commit
moves the interface to where it belongs -- in `pkg/cmd/pulumi`,
factoring related backend code to a new `backend` submodule while we're
at it. This simplifies `pkg/backend` (and sets us up for more clean-up
later on) and pulls more gunk out of `pkg/cmd/pulumi/util.go` to boot.