Remove second go.mod in parameterized packages (#17928)
This code wrote a go.mod file one folder below where we actually
expected a go.mod file to be written according to SupportPack, since
https://github.com/pulumi/pulumi/pull/17923 was merged we were writing
out both. This removes the nested go.mod, fixes the integration test
that was using that instead of the top level go.mod, and fixes `package
add` to refer to the top level go.mod.