[backend/diy] Show a more correct URL for lock files
In the DIY backend, when a stack is locked, we attempt to show the URL to the lock file in the error message we show to users. The URL is naively created by concatenating the baseURL with the path to the lock file. But this naive concatenation isn't correct when there are query params in the base URL. This change fixes that by concatenating the path of a parsed URL. If for some reason we can't parse the URL, we fall back to the naive concatenation we were doing previously.