Add ImportIfNew as resource option (#3388)
This flag is designed to change the behaviour of the deployment engine
so that the ImportID of a resource is updated, but only used to generate
an import step if the resource is not yet part of the tracked state.

This makes it much easier to write Pulumi programs which search for existing
resources in an environment and set the import ID accordingly.

Previously, setting the import ID of a resource would always perform a
comparison with the previous ID and import ID. This caused a problem with
resource types (usually ones with impromptu composite IDs) that, during
the import, would get a completely different ID generated by the
provider. ImportIfNew allows this mess to be side-stepped.