Modify packages for the user instead of printing instructions when adding package. (#17815)
This is related to the the previous attempt #17795 where we discovered
the actual root issue wasn't the paths but that a sub package was being
generated unnecessarily (and the paths were ok).

If there are no local dependencies that indicates this is actually
needed and not a duplicate go.mod (differentiating between adding a
package and converting a project etc).

This also contains #17814 which merged into it:

Before we would print instructions, but with built in tools it is not
hard to modify the go.mod on behalf of the user and print out
information.

This still leaves java in a "print instruction" state since there isnt a
standard package managment system like in the other languages.

Fixes https://github.com/pulumi/pulumi/issues/17902

---------

Co-authored-by: Thomas Gummerer <t.gummerer@gmail.com>