[sdk/nodejs] Add getSchema to Provider interface (#17950)
Originally implemented in https://github.com/pulumi/pulumi/pull/6892,
node's Provider interface has a `schema` property of type string. As
such, it's required to be set unconditionally at provider class
instantiation, not loaded dynamically on demand. This puts an unneeded
burden on provider startup to load or calculate the entire schema even
though it may never be needed.

This PR adds a `getSchema?: () => Promise<string>;` method in line with
gRPC interface functionality. `schema` is still present for backwards
compatibility. The server will use `schema` if it's defined, or
`getSchema` if not.